made some changes

+++ programming_paradigm/bank_account.py
@@ -3,12 +3,12 @@ class BankAccount:
         self.account_balance = 0
 
     def deposit(self,amount):
-        amount = int(input('Amount to be deposited: '))
+        amount = float(input('Amount to be deposited: '))
         self.account_balance = self.account_balance+amount
         print(f'Deposited: ${amount}')
 
     def withdraw(self,amount):
-        amount = int(input('Amount to withdraw: '))
+        amount = float(input('Amount to withdraw: '))
         if self.account_balance >= amount:
             self.account_balance = self.account_balance - amount
             print(f'Amount withdrawn: ${amount}')
